(PUP-9570) Accept check_environment parameter
Accept the parameter and include it with the v3 catalog request. By default, the
value is false, as there may be other users of the http client API that aren't
expecting the new behavior. The double negation is needed in case `nil` is
passed and behaves the same as `false`.